Allow external traffic only to img.youtube.com
This uses sniproxy to allow selective, whitelisted external
HTTPS access only. sniproxy will do a DNS request for the
domain name in the clienthello and connect only to that
domain. This *should* protect us from arbitrary data
exfiltration, since the attacker can not control the
IP that is being connected to. Domain fronting can still
get over this, but img.youtube.com can not be used as a
Domain front afaict, so we're ok.